+++ QuickTableViewControllerTests/TapActionRowSpec.swift
@@ -37,6 +37,7 @@ class TapActionRowSpec: QuickSpec {
 
       it("should initialize with given parameters") {
         expect(row.title) == "title"
+        expect(row.cellReuseIdentifier) == "TapActionCell"
         expect(row.action).notTo(beNil())
 
         row.action?(row)